requirements.txt
# fastapi
# uvicorn[standard]
# pydantic

# --- ML/Data ---
scikit-learn
numpy
nltk
pyyaml
joblib

# --- Deep Learning (GRU model) ---
torch

# --- API ---
fastapi
uvicorn[standard]
pydantic
